Ajout funding controller

+++ ohana-project/src/main/java/fr/isika/cda11/ohana/project/crowdfunding/service/FundingService.java
@@ -34,7 +34,7 @@ public class FundingService {
 		
 	}
 	
-	public Funding findFundingService(Integer id) {
+	public Funding findFundingService(Long id) {
 		
 		return fundingRepos.findFundingRepos(id);
